requirements.txt
PySide6>=6.6.1
psutil==5.9.8
requests==2.31.0
pyyaml>=6.0
docker==7.0.0
tqdm>=4.65.0
aiohttp>=3.8.0
asyncio>=3.4.3
python-dotenv>=0.19.0
pyinstaller>=6.5.0 